+++ packages/xlsx-preview/src/schema/index.ts
@@ -22,6 +22,7 @@ export type { DefinedName } from "./DefinedName.js";
 export type { Drawing } from "./Drawing.js";
 export type { Dxf } from "./Dxf.js";
 export type { DrawingAnchor } from "./DrawingAnchor.js";
+export type { DrawingHyperlink } from "./DrawingHyperlink.js";
 export type { Fill } from "./Fill.js";
 export type { GradientStop } from "./GradientStop.js";
 export type { Font } from "./Font.js";